Layout seed: exact-text numerics + published md5, for the migration's guard
Migration 0010 refuses to apply if the seed it carries has been hand-edited.
That guard is only worth having if it compares hashes that provably agree, and
as written it would not have.

Two changes make the hash reproducible across Node and Postgres:

  * the seed's temp table declares relative_x/relative_y/absolute_lat/lng as
    NUMERIC rather than double precision. numeric round-trips through ::text
    exactly ('10.0000' stays '10.0000'); a double can print as
    141.28940000000001, which would make the same data hash differently
    depending on who formatted it. The migration casts to double precision on
    the way into public.stalls, so the stored type is unchanged.

  * the generator now computes and publishes the md5 the migration checks, over
    stall_code|stall_type|x|y|heading|width|depth, newline-joined, ordered by
    stall_code.

Both halves verified against the live server rather than assumed:
  * numeric::text formatting confirmed on Postgres (10.0000, 0.0000, 452.1300)
  * the ordering confirmed under the database's actual en_US.UTF-8 collation --
    a locale that de-prioritises punctuation could easily have sorted these
    hyphenated codes differently from JavaScript's code-unit sort, and that
    would have aborted a perfectly good migration. Postgres and the generator
    return the identical hash over all 160 codes.

Seed md5: 89943752f32f98da4ef80b6baedb3174

+++ unreal/layoutSeed.sql
@@ -12,6 +12,11 @@
 -- Structures: 26
 -- Retired stall codes:     19
 -- Retired structure codes: 2  (CANOPY-04, METAL-CANOPY-PERIM)
+--
+-- SEED MD5: 89943752f32f98da4ef80b6baedb3174
+--   md5 over stall_code|stall_type|relative_x|relative_y|heading|width|depth,
+--   newline-joined, ordered by stall_code. Migration 0010 recomputes this in SQL
+--   and aborts on mismatch, so a hand-edited seed cannot reshape the depot.
 -- =============================================================================
 
 -- The seed is DEPOT-AGNOSTIC. stall_code is the natural key; the migration
@@ -26,13 +31,13 @@ CREATE TEMP TABLE ottoq_layout_seed_stalls (
   zone             text        NOT NULL,
   staging_role     text,
   display_name     text        NOT NULL,
-  relative_x       double precision NOT NULL,
-  relative_y       double precision NOT NULL,
+  relative_x       numeric     NOT NULL,
+  relative_y       numeric     NOT NULL,
   heading_degrees  smallint    NOT NULL,
   stall_width_ft   numeric     NOT NULL,
   stall_depth_ft   numeric     NOT NULL,
-  absolute_lat     double precision NOT NULL,
-  absolute_lng     double precision NOT NULL,
+  absolute_lat     numeric     NOT NULL,
+  absolute_lng     numeric     NOT NULL,
   canopy_code      text,
   canopy_side      text,
   covered          boolean     NOT NULL,
